refactor: modernize C++20 style and streamline system metadata
Prefer ranges/contains/const-ref patterns with teaching comments, inline
GoogleTest setup into the root CMakeLists, and extract db.meta JSON
handling into sm_meta.cpp with focused unit coverage.

+++ src/execution/executor_nestedloop_join.h
@@ -9,44 +9,38 @@
 #include "system/sm.h"
 
 class NestedLoopJoinExecutor : public AbstractExecutor {
-   private:
-    std::unique_ptr<AbstractExecutor> left_;    // 左儿子节点（需要join的表）
-    std::unique_ptr<AbstractExecutor> right_;   // 右儿子节点（需要join的表）
-    size_t len_;                                // join后获得的每条记录的长度
-    std::vector<ColMeta> cols_;                 // join后获得的记录的字段
+private:
+    std::unique_ptr<AbstractExecutor> left_;   // 左儿子节点（需要join的表）
+    std::unique_ptr<AbstractExecutor> right_;  // 右儿子节点（需要join的表）
+    size_t len_;                               // join后获得的每条记录的长度
+    std::vector<ColMeta> cols_;                // join后获得的记录的字段
 
-    std::vector<Condition> fed_conds_;          // join条件
+    std::vector<Condition> fed_conds_;  // join条件
     bool isend;
 
-   public:
-    NestedLoopJoinExecutor(std::unique_ptr<AbstractExecutor> left, std::unique_ptr<AbstractExecutor> right, 
-                            std::vector<Condition> conds) {
+public:
+    NestedLoopJoinExecutor(std::unique_ptr<AbstractExecutor> left,
+                           std::unique_ptr<AbstractExecutor> right,
+                           std::vector<Condition> conds) {
         left_ = std::move(left);
         right_ = std::move(right);
         len_ = left_->tupleLen() + right_->tupleLen();
         cols_ = left_->cols();
         auto right_cols = right_->cols();
-        for (auto &col : right_cols) {
-            col.offset += left_->tupleLen();
+        for (auto& col : right_cols) {
+            col.offset += static_cast<int>(left_->tupleLen());
         }
 
         cols_.insert(cols_.end(), right_cols.begin(), right_cols.end());
         isend = false;
         fed_conds_ = std::move(conds);
-
     }
 
-    void beginTuple() override {
-        throw NotImplementedError("NestedLoopJoinExecutor::beginTuple");
-    }
+    void beginTuple() override { throw NotImplementedError("NestedLoopJoinExecutor::beginTuple"); }
 
-    void nextTuple() override {
-        throw NotImplementedError("NestedLoopJoinExecutor::nextTuple");
-    }
+    void nextTuple() override { throw NotImplementedError("NestedLoopJoinExecutor::nextTuple"); }
 
-    std::unique_ptr<RmRecord> Next() override {
-        throw NotImplementedError("NestedLoopJoinExecutor::Next");
-    }
+    std::unique_ptr<RmRecord> Next() override { throw NotImplementedError("NestedLoopJoinExecutor::Next"); }
 
-    Rid &rid() override { return _abstract_rid; }
+    Rid& rid() override { return _abstract_rid; }
 };
