+++ src/execution/executor_projection.h
@@ -9,19 +9,19 @@
 #include "system/sm.h"
 
 class ProjectionExecutor : public AbstractExecutor {
-   private:
-    std::unique_ptr<AbstractExecutor> prev_;        // 投影节点的儿子节点
-    std::vector<ColMeta> cols_;                     // 需要投影的字段
-    size_t len_;                                    // 字段总长度
-    std::vector<size_t> sel_idxs_;                  
+private:
+    std::unique_ptr<AbstractExecutor> prev_;  // 投影节点的儿子节点
+    std::vector<ColMeta> cols_;               // 需要投影的字段
+    size_t len_;                              // 字段总长度
+    std::vector<size_t> sel_idxs_;
 
-   public:
-    ProjectionExecutor(std::unique_ptr<AbstractExecutor> prev, const std::vector<TabCol> &sel_cols) {
+public:
+    ProjectionExecutor(std::unique_ptr<AbstractExecutor> prev, const std::vector<TabCol>& sel_cols) {
         prev_ = std::move(prev);
 
-        size_t curr_offset = 0;
-        auto &prev_cols = prev_->cols();
-        for (auto &sel_col : sel_cols) {
+        int curr_offset = 0;
+        auto& prev_cols = prev_->cols();
+        for (auto& sel_col : sel_cols) {
             auto pos = get_col(prev_cols, sel_col);
             sel_idxs_.push_back(pos - prev_cols.begin());
             auto col = *pos;
@@ -36,9 +36,7 @@ class ProjectionExecutor : public AbstractExecutor {
 
     void nextTuple() override { throw NotImplementedError("ProjectionExecutor::nextTuple"); }
 
-    std::unique_ptr<RmRecord> Next() override {
-        throw NotImplementedError("ProjectionExecutor::Next");
-    }
+    std::unique_ptr<RmRecord> Next() override { throw NotImplementedError("ProjectionExecutor::Next"); }
 
-    Rid &rid() override { return _abstract_rid; }
+    Rid& rid() override { return _abstract_rid; }
 };
